requirements.txt
aiogram==3.10.0
python-dotenv==1.0.0
motor==3.3.2
pymongo==4.6.0
openai==1.14.0
asyncio
logging 